refactor: address CodeRabbit module-size and VRAM review notes
Split Gradio pipeline CLI/startup into focused modules under the 200 LOC
cap, extract device-resolution tests, add missing test helper docstrings,
and read LM free/total VRAM from the mapped cuda:N device.
