+++ acestep/acestep_v15_pipeline_test.py
@@ -37,7 +37,7 @@ class PipelineStartupBackendTests(unittest.TestCase):
         argv: list[str],
         *,
         env: dict[str, str] | None = None,
-    ) -> tuple[MagicMock, dict[str, object]]:
+    ) -> tuple[MagicMock, MagicMock, dict[str, object]]:
         """Run ``main`` with heavy dependencies stubbed and capture startup state."""
         gpu_config = self._legacy_gpu_config()
         dit_handler = MagicMock()
@@ -52,15 +52,17 @@ class PipelineStartupBackendTests(unittest.TestCase):
         demo = MagicMock()
         demo.queue.return_value = demo
         demo.launch.return_value = None
-
         captured: dict[str, object] = {}
 
         def _create_demo(init_params=None, language="en"):
+            """Capture init_params while returning a stub Gradio demo."""
             captured["init_params"] = init_params
             captured["language"] = language
             return demo
 
-        with patch.object(sys, "argv", argv), patch.dict(os.environ, env or {}, clear=True), patch(
+        with patch.object(sys, "argv", argv), patch.dict(
+            os.environ, env or {}, clear=True
+        ), patch(
             "acestep.acestep_v15_pipeline.get_gpu_config",
             return_value=gpu_config,
         ), patch(
@@ -71,19 +73,19 @@ class PipelineStartupBackendTests(unittest.TestCase):
         ), patch(
             "acestep.acestep_v15_pipeline.get_i18n"
         ), patch(
-            "acestep.acestep_v15_pipeline.available_languages_info",
+            "acestep.gradio_pipeline_cli.available_languages_info",
             return_value=[("en", "English", "English")],
         ), patch(
-            "acestep.acestep_v15_pipeline.AceStepHandler",
+            "acestep.gradio_pipeline_startup.AceStepHandler",
             return_value=dit_handler,
         ), patch(
-            "acestep.acestep_v15_pipeline.LLMHandler",
+            "acestep.gradio_pipeline_startup.LLMHandler",
             return_value=llm_handler,
         ), patch(
             "acestep.acestep_v15_pipeline.create_demo",
             side_effect=_create_demo,
         ), patch(
-            "acestep.acestep_v15_pipeline.ensure_lm_model",
+            "acestep.gradio_pipeline_startup.ensure_lm_model",
             return_value=(True, "ok"),
         ), patch(
             "acestep.acestep_v15_pipeline.os.makedirs"
@@ -109,7 +111,6 @@ class PipelineStartupBackendTests(unittest.TestCase):
                 "vllm",
             ]
         )
-
         self.assertEqual("pt", llm_handler.initialize.call_args.kwargs["backend"])
         self.assertEqual("pt", captured["init_params"]["backend"])
 
@@ -119,14 +120,12 @@ class PipelineStartupBackendTests(unittest.TestCase):
             ["acestep", "--service_mode", "true", "--init_llm", "true"],
             env={"SERVICE_MODE_BACKEND": "vllm"},
         )
-
         self.assertEqual("pt", llm_handler.initialize.call_args.kwargs["backend"])
         self.assertEqual("pt", captured["init_params"]["backend"])
 
     def test_main_forces_pt_backend_for_api_env_override(self) -> None:
         """API-mode env overrides should still resolve to the safe startup backend."""
         api_routes_module = types.SimpleNamespace(setup_api_routes=MagicMock())
-
         with patch.dict(
             sys.modules,
             {"acestep.ui.gradio.api.api_routes": api_routes_module},
@@ -144,113 +143,8 @@ class PipelineStartupBackendTests(unittest.TestCase):
                 ],
                 env={"ACESTEP_LM_BACKEND": "vllm"},
             )
-
         self.assertEqual("pt", llm_handler.initialize.call_args.kwargs["backend"])
         self.assertEqual("pt", captured["init_params"]["backend"])
-
-
-class PipelineGpuMappingTests(unittest.TestCase):
-    """Verify multi-GPU CLI flags are wired into service initialization."""
-
-    def test_list_gpus_exits_after_printing_inventory(self) -> None:
-        with patch.object(sys, "argv", ["acestep", "--list-gpus"]), patch(
-            "acestep.acestep_v15_pipeline.format_gpu_list_text",
-            return_value="GPU TABLE",
-        ) as mock_format, patch(
-            "acestep.acestep_v15_pipeline.sys.exit",
-            side_effect=SystemExit(0),
-        ) as mock_exit:
-            with self.assertRaises(SystemExit):
-                acestep_v15_pipeline.main()
-        mock_format.assert_called_once()
-        mock_exit.assert_called_once_with(0)
-
-    def test_gpu_mapping_passed_to_initialize_service(self) -> None:
-        gpu_config = SimpleNamespace(
-            gpu_memory_gb=24.0,
-            tier="tier6b",
-            max_duration_with_lm=480,
-            max_duration_without_lm=600,
-            max_batch_size_with_lm=8,
-            max_batch_size_without_lm=8,
-            init_lm_default=True,
-            available_lm_models=["acestep-5Hz-lm-0.6B"],
-            recommended_backend="vllm",
-            lm_backend_restriction=None,
-            offload_dit_to_cpu_default=False,
-            quantization_default=False,
-        )
-        dit_handler = MagicMock()
-        dit_handler.get_available_acestep_v15_models.return_value = ["acestep-v15-turbo"]
-        dit_handler.is_flash_attention_available.return_value = False
-        dit_handler.initialize_service.return_value = ("ok", True)
-        dit_handler.device_map = SimpleNamespace(lm="cuda:1")
-
-        llm_handler = MagicMock()
-        llm_handler.get_available_5hz_lm_models.return_value = ["acestep-5Hz-lm-0.6B"]
-        llm_handler.initialize.return_value = ("ok", True)
-
-        demo = MagicMock()
-        demo.queue.return_value = demo
-        demo.launch.return_value = None
-        captured: dict[str, object] = {}
-
-        def _create_demo(init_params=None, language="en"):
-            captured["init_params"] = init_params
-            return demo
-
-        with patch.object(
-            sys,
-            "argv",
-            [
-                "acestep",
-                "--init_service",
-                "true",
-                "--init_llm",
-                "true",
-                "--config_path",
-                "acestep-v15-turbo",
-                "--gpu-mapping",
-                "auto",
-            ],
-        ), patch.dict(os.environ, {}, clear=True), patch(
-            "acestep.acestep_v15_pipeline.get_gpu_config",
-            return_value=gpu_config,
-        ), patch(
-            "acestep.acestep_v15_pipeline.set_global_gpu_config"
-        ), patch(
-            "acestep.acestep_v15_pipeline.is_mps_platform",
-            return_value=False,
-        ), patch(
-            "acestep.acestep_v15_pipeline.get_i18n"
-        ), patch(
-            "acestep.acestep_v15_pipeline.available_languages_info",
-            return_value=[("en", "English", "English")],
-        ), patch(
-            "acestep.acestep_v15_pipeline.AceStepHandler",
-            return_value=dit_handler,
-        ), patch(
-            "acestep.acestep_v15_pipeline.LLMHandler",
-            return_value=llm_handler,
-        ), patch(
-            "acestep.acestep_v15_pipeline.create_demo",
-            side_effect=_create_demo,
-        ), patch(
-            "acestep.acestep_v15_pipeline.ensure_lm_model",
-            return_value=(True, "ok"),
-        ), patch(
-            "acestep.acestep_v15_pipeline.os.makedirs"
-        ), patch(
-            "acestep.acestep_v15_pipeline.log_lm_device_deprecation"
-        ):
-            acestep_v15_pipeline.main()
-
-        self.assertEqual(
-            "auto",
-            dit_handler.initialize_service.call_args.kwargs["gpu_mapping"],
-        )
-        self.assertEqual("cuda:1", llm_handler.initialize.call_args.kwargs["device"])
-        self.assertEqual("auto", captured["init_params"]["gpu_mapping"])
 
 
 if __name__ == "__main__":
